+++ components/ProfilePage/Deploy6551AndMintButton.tsx
@@ -1,11 +1,9 @@
-import { Contract } from "ethers"
 import { FC, useState } from "react"
 import { Button } from "../../shared/Button"
-import abi from "../../lib/abi-dna-minter.json"
 import { useProfileProvider } from "../../providers/ProfileContext"
-import { useEthersSigner } from "../../hooks/useEthersSigner"
 import SettingSmartWalletModal from "./SettingSmartWalletModal"
 import useCheckNetwork from "../../hooks/useCheckNetwork"
+import useCreateTBA from "../../hooks/useCreateTBA"
 
 interface Deploy6551AndMintDNAButtonProps {
   getDNAByCre8orNumber: any
@@ -16,18 +14,14 @@ const Deploy6551AndMintDNAButton: FC<Deploy6551AndMintDNAButtonProps> = ({
 }) => {
   const { cre8orNumber } = useProfileProvider()
   const { checkNetwork } = useCheckNetwork()
-
-  const signer = useEthersSigner({ chainId: process.env.NEXT_PUBLIC_TESTNET ? 5 : 1 })
+  const { createTbaAndMintDna } = useCreateTBA()
   const [openLoadingModal, setOpenLoadingModal] = useState(false)
 
   const onClick = async () => {
     if (!checkNetwork()) return
 
     setOpenLoadingModal(true)
-    const dnaMinter = process.env.NEXT_PUBLIC_DNA_MINTER
-    const contract = new Contract(dnaMinter, abi, signer)
-    const tx = await contract.createTokenBoundAccountAndMintDNA(cre8orNumber)
-    await tx.wait()
+    await createTbaAndMintDna(cre8orNumber)
     getDNAByCre8orNumber()
     setOpenLoadingModal(false)
   }
